resoliq_api-main/api/requirements.txt
django
djangorestframework
psycopg2-binary
gunicorn
Pillow
django-cors-headers
django-filter
requests
django-import-export
